Refactor operations
